Migrate existing scope tests to new path
