--- src/globus_sdk/experimental/scope_parser/errors.py
@@ -1,6 +1,0 @@
-class ScopeParseError(ValueError):
-    """The error raised if scope parsing fails."""
-
-
-class ScopeCycleError(ScopeParseError):
-    """The error raised if scope parsing discovers a cycle."""
